refactor: move manifest generation from pkg/cargo to pkg/manifest
