+++ pkg/manifest/fixtures/manifest.yml
