+++ pkg/manifest/fixtures/metadata.yml
