+++ pkg/manifest/opsman/resource_config.go
