+++ pkg/manifest/opsman/stemcell.go
